Boss Changes
Changes boss movement so it stays near the top of the screen. Boss also fires a projectile randomly every so often. Tried adding health to the boss, but am having trouble figuring out how the enemy_4 damage is handled. I only seem able to change the health of the shields and not of the enemy itself. Boss projectiles do not yet damage the hero.