Health Bar
Finally have a working health bar. I was initially making a reference to the enemy_4 prefab instead of the instance that is instantiated in the game. Health bar game object is destroyed when the boss is defeated. Boss will now only be added to the list of enemies to be spawned once 15 enemies have been spawned to give player time to get shield and powerups