Fixed Health Bar and Boss Enemy
Originally once the boss was defeated I was just destroying the canvas gameobject that the health bar was in. Because the text that displays after the boss is defeated is a child of the same canvas, that was also getting deleted. Now both the health bar and the winText  just enable and disable using gameObject.SetActive(). Because of the way the enemy class is setup, if you ran into the boss enemy with shield around the hero, the boss would instantly die. I thought this might be fun to leave in to see if people would figure it out, but I decided to remove it.